bot_app/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
distro==1.9.0
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
nats-py==2.9.0
netaddr==1.3.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
redis==5.2.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
yookassa==3.5.0
openpyxl==3.1.5
